fix: move tests outside package, fix test CI failure
This change moves all the test code and test data out of the Python
package, which means that it no longer takes up space in the built
distribution, and is no longer installed with the package on end-user
machines.

It also fixes the CI script that assumed tests ran inside the package,
which depended upon the local imports that were removed in the last
batch of changes.